refactor: move another types file to the root
This types file does not need to be in a 'backend'
folder. Instead it is moved to the root of the webview
code.
